clamp the number of leading zeros
The test data provided create a vdelta that has 33 leading bits.  This
value doesn't fit within the 5 bits provided for storing this length,
and the floats cannot be round-tripped.

By pretending we have fewer leading zeros, we avoid this problem.
Clamping is also nicer than increasing the number of bits to 6 (even
though this has a negligible affect on production data sizes.)

Fixes #4